use a state and move things around
