+++ source/s3_client.c
@@ -1885,13 +1885,14 @@ static void s_s3_client_body_streaming_task(struct aws_task *task, void *arg, en
         struct aws_s3_request *request = AWS_CONTAINER_OF(request_node, struct aws_s3_request, node);
         struct aws_s3_meta_request *meta_request = request->meta_request;
 
-        bool cancelled = aws_atomic_load_int(&meta_request->cancelled) == 1;
+        aws_s3_meta_request_lock_synced_data(meta_request);
+        bool active = meta_request->synced_data.state == AWS_S3_META_REQUEST_STATE_ACTIVE;
+        aws_s3_meta_request_unlock_synced_data(meta_request);
 
-        if (cancelled) {
-            /* meta request has been cancelled, drop the body after that */
+        if (!active) {
             AWS_LOGF_DEBUG(
                 AWS_LS_S3_CLIENT,
-                "id=%p meta request %p cancelled, drop the body.",
+                "id=%p meta request %p is not active, drop the body.",
                 (void *)client,
                 (void *)meta_request);
             aws_s3_request_release(request);
