+++ source/s3_meta_request.c
@@ -208,20 +208,20 @@ int aws_s3_meta_request_init_base(
     meta_request->finish_callback = options->finish_callback;
     meta_request->shutdown_callback = options->shutdown_callback;
 
-    aws_atomic_init_int(&meta_request->cancelled, 0);
-
     return AWS_OP_SUCCESS;
 }
 
 void aws_s3_meta_request_cancel(struct aws_s3_meta_request *meta_request) {
     AWS_PRECONDITION(meta_request);
     AWS_PRECONDITION(meta_request->vtable->finish);
-    bool cancel_called = true;
-    if (aws_atomic_load_int(&meta_request->cancelled) == 0) {
-        cancel_called = false;
-        aws_atomic_store_int(&meta_request->cancelled, 1);
+
+    aws_s3_meta_request_lock_synced_data(meta_request);
+    bool active = meta_request->synced_data.state == AWS_S3_META_REQUEST_STATE_ACTIVE;
+    if (active) {
+        meta_request->synced_data.state = AWS_S3_META_REQUEST_STATE_CANCELLING;
     }
-    if (!cancel_called) {
+    aws_s3_meta_request_unlock_synced_data(meta_request);
+    if (active) {
         meta_request->vtable->finish(meta_request, NULL, AWS_ERROR_S3_CANCELED);
     }
 }
@@ -888,7 +888,6 @@ void aws_s3_meta_request_finish_default(
     struct aws_s3_request *failed_request,
     int error_code) {
 
-    aws_atomic_store_int(&meta_request->cancelled, 1);
     if (failed_request != NULL) {
         aws_s3_meta_request_finish(meta_request, failed_request, failed_request->send_data.response_status, error_code);
     } else {
